protegendo end-points

+++ src/tasks/tasks.service.ts
@@ -3,6 +3,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { PaginationDto } from '../common/dto/pagination.dto';
+import { PayloadTokenDto } from '../auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -35,14 +36,14 @@ export class TasksService {
     // throw new NotFoundException('Essa tarefa NAO EXISTE!');
   }
 
-  async create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const newTask = await this.prisma.task.create({
         data: {
           name: createTaskDto.name,
           description: createTaskDto.description,
           completed: false,
-          userId: createTaskDto.userId,
+          userId: tokenPayload.sub,
         },
       });
 
@@ -56,7 +57,11 @@ export class TasksService {
     }
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto) {
+  async update(
+    id: number,
+    updateTaskDto: UpdateTaskDto,
+    tokenPayload: PayloadTokenDto,
+  ) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
@@ -65,6 +70,13 @@ export class TasksService {
       });
 
       if (!findTask) {
+        throw new HttpException(
+          'Essa tarefa não existe',
+          HttpStatus.BAD_REQUEST,
+        );
+      }
+
+      if (findTask.userId !== tokenPayload.sub) {
         throw new HttpException(
           'Essa tarefa não existe',
           HttpStatus.BAD_REQUEST,
@@ -95,7 +107,7 @@ export class TasksService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
@@ -105,6 +117,13 @@ export class TasksService {
 
       if (!findTask) {
         throw new HttpException('Essa tarefa não existe', HttpStatus.NOT_FOUND);
+      }
+
+      if (findTask.userId !== tokenPayload.sub) {
+        throw new HttpException(
+          'Falha ao deletar essa tarefa!',
+          HttpStatus.BAD_REQUEST,
+        );
       }
 
       await this.prisma.task.delete({
